Move Gh functions into new files, GeneralizedHarmonic namespace.
