+++ src/PointwiseFunctions/GeneralRelativity/ComputeGhQuantities.cpp
@@ -1,0 +1,121 @@
+// Distributed under the MIT License.
+// See LICENSE.txt for details.
+
+#include "PointwiseFunctions/GeneralRelativity/ComputeGhQuantities.hpp"
+
+#include "DataStructures/DataVector.hpp"
+#include "DataStructures/MakeWithValue.hpp"
+#include "DataStructures/Tensor/Tensor.hpp"
+#include "Utilities/GenerateInstantiations.hpp"
+
+namespace GeneralizedHarmonic {
+template <size_t SpatialDim, typename Frame, typename DataType>
+tnsr::iaa<DataType, SpatialDim, Frame> phi(
+    const Scalar<DataType>& lapse,
+    const tnsr::i<DataType, SpatialDim, Frame>& deriv_lapse,
+    const tnsr::I<DataType, SpatialDim, Frame>& shift,
+    const tnsr::iJ<DataType, SpatialDim, Frame>& deriv_shift,
+    const tnsr::ii<DataType, SpatialDim, Frame>& spatial_metric,
+    const tnsr::ijj<DataType, SpatialDim, Frame>&
+        deriv_spatial_metric) noexcept {
+  tnsr::iaa<DataType, SpatialDim, Frame> phi(
+      make_with_value<DataType>(deriv_lapse, 0.));
+  for (size_t k = 0; k < SpatialDim; ++k) {
+    phi.get(k, 0, 0) = -2.0 * lapse.get() * deriv_lapse.get(k);
+    for (size_t m = 0; m < SpatialDim; ++m) {
+      for (size_t n = 0; n < SpatialDim; ++n) {
+        phi.get(k, 0, 0) +=
+            deriv_spatial_metric.get(k, m, n) * shift.get(m) * shift.get(n) +
+            2.0 * spatial_metric.get(m, n) * shift.get(m) *
+                deriv_shift.get(k, n);
+      }
+    }
+
+    for (size_t i = 0; i < SpatialDim; ++i) {
+      for (size_t m = 0; m < SpatialDim; ++m) {
+        phi.get(k, 0, i + 1) +=
+            deriv_spatial_metric.get(k, m, i) * shift.get(m) +
+            spatial_metric.get(m, i) * deriv_shift.get(k, m);
+      }
+      for (size_t j = i; j < SpatialDim; ++j) {
+        phi.get(k, i + 1, j + 1) = deriv_spatial_metric.get(k, i, j);
+      }
+    }
+  }
+  return phi;
+}
+
+template <size_t SpatialDim, typename Frame, typename DataType>
+tnsr::aa<DataType, SpatialDim, Frame> pi(
+    const Scalar<DataType>& lapse, const Scalar<DataType>& dt_lapse,
+    const tnsr::I<DataType, SpatialDim, Frame>& shift,
+    const tnsr::I<DataType, SpatialDim, Frame>& dt_shift,
+    const tnsr::ii<DataType, SpatialDim, Frame>& spatial_metric,
+    const tnsr::ii<DataType, SpatialDim, Frame>& dt_spatial_metric,
+    const tnsr::iaa<DataType, SpatialDim, Frame>& phi) noexcept {
+  tnsr::aa<DataType, SpatialDim, Frame> pi{
+      make_with_value<DataType>(lapse, 0.)};
+
+  pi.get(0, 0) = -2.0 * lapse.get() * dt_lapse.get();
+
+  for (size_t m = 0; m < SpatialDim; ++m) {
+    for (size_t n = 0; n < SpatialDim; ++n) {
+      pi.get(0, 0) +=
+          dt_spatial_metric.get(m, n) * shift.get(m) * shift.get(n) +
+          2.0 * spatial_metric.get(m, n) * shift.get(m) * dt_shift.get(n);
+    }
+  }
+
+  for (size_t i = 0; i < SpatialDim; ++i) {
+    for (size_t m = 0; m < SpatialDim; ++m) {
+      pi.get(0, i + 1) += dt_spatial_metric.get(m, i) * shift.get(m) +
+                          spatial_metric.get(m, i) * dt_shift.get(m);
+    }
+    for (size_t j = i; j < SpatialDim; ++j) {
+      pi.get(i + 1, j + 1) = dt_spatial_metric.get(i, j);
+    }
+  }
+  for (size_t mu = 0; mu < SpatialDim + 1; ++mu) {
+    for (size_t nu = mu; nu < SpatialDim + 1; ++nu) {
+      for (size_t i = 0; i < SpatialDim; ++i) {
+        pi.get(mu, nu) -= shift.get(i) * phi.get(i, mu, nu);
+      }
+      pi.get(mu, nu) /= -lapse.get();
+    }
+  }
+  return pi;
+}
+}  // namespace GeneralizedHarmonic
+
+/// \cond
+#define DIM(data) BOOST_PP_TUPLE_ELEM(0, data)
+#define DTYPE(data) BOOST_PP_TUPLE_ELEM(1, data)
+#define FRAME(data) BOOST_PP_TUPLE_ELEM(2, data)
+
+#define INSTANTIATE(_, data)                                                  \
+  template tnsr::iaa<DTYPE(data), DIM(data), FRAME(data)>                     \
+  GeneralizedHarmonic::phi(                                                   \
+      const Scalar<DTYPE(data)>& lapse,                                       \
+      const tnsr::i<DTYPE(data), DIM(data), FRAME(data)>& deriv_lapse,        \
+      const tnsr::I<DTYPE(data), DIM(data), FRAME(data)>& shift,              \
+      const tnsr::iJ<DTYPE(data), DIM(data), FRAME(data)>& deriv_shift,       \
+      const tnsr::ii<DTYPE(data), DIM(data), FRAME(data)>& spatial_metric,    \
+      const tnsr::ijj<DTYPE(data), DIM(data), FRAME(data)>&                   \
+          deriv_spatial_metric) noexcept;                                     \
+  template tnsr::aa<DTYPE(data), DIM(data), FRAME(data)>                      \
+  GeneralizedHarmonic::pi(                                                    \
+      const Scalar<DTYPE(data)>& lapse, const Scalar<DTYPE(data)>& dt_lapse,  \
+      const tnsr::I<DTYPE(data), DIM(data), FRAME(data)>& shift,              \
+      const tnsr::I<DTYPE(data), DIM(data), FRAME(data)>& dt_shift,           \
+      const tnsr::ii<DTYPE(data), DIM(data), FRAME(data)>& spatial_metric,    \
+      const tnsr::ii<DTYPE(data), DIM(data), FRAME(data)>& dt_spatial_metric, \
+      const tnsr::iaa<DTYPE(data), DIM(data), FRAME(data)>& phi) noexcept;
+
+GENERATE_INSTANTIATIONS(INSTANTIATE, (1, 2, 3), (double, DataVector),
+                        (Frame::Grid, Frame::Inertial))
+
+#undef DIM
+#undef DTYPE
+#undef FRAME
+#undef INSTANTIATE
+/// \endcond
