+++ tests/Unit/PointwiseFunctions/GeneralRelativity/CMakeLists.txt
@@ -4,6 +4,7 @@
 set(SPECTRE_UNIT_GENERAL_RELATIVITY
     PointwiseFunctions/GeneralRelativity/GrTestHelpers.cpp
     PointwiseFunctions/GeneralRelativity/Test_Christoffel.cpp
+    PointwiseFunctions/GeneralRelativity/Test_ComputeGhQuantities.cpp
     PointwiseFunctions/GeneralRelativity/Test_ComputeSpacetimeQuantities.cpp
     PointwiseFunctions/GeneralRelativity/Test_IndexManipulation.cpp
     )
